fix Ruslan's comments
