+++ cpp/oneapi/dal/algo/pca/backend/cpu/infer_kernel.hpp
@@ -21,7 +21,7 @@
 
 namespace oneapi::dal::pca::backend {
 
-template <typename Float, typename Method, typename Task>
+template <typename Float, typename Task>
 struct infer_kernel_cpu {
     infer_result<Task> operator()(const dal::backend::context_cpu& ctx,
                                   const descriptor_base<Task>& params,
