+++ cpp/oneapi/dal/algo/pca/backend/gpu/infer_kernel.hpp
@@ -21,7 +21,7 @@
 
 namespace oneapi::dal::pca::backend {
 
-template <typename Float, typename Method, typename Task>
+template <typename Float, typename Task>
 struct infer_kernel_gpu {
     infer_result<Task> operator()(const dal::backend::context_gpu& ctx,
                                   const descriptor_base<Task>& params,
