+++ cpp/oneapi/dal/algo/pca/detail/infer_ops_dpc.cpp
@@ -28,8 +28,8 @@ struct ONEAPI_DAL_EXPORT infer_ops_dispatcher<data_parallel_policy, Float, Metho
                                   const descriptor_base<Task>& params,
                                   const infer_input<Task>& input) const {
         using kernel_dispatcher_t =
-            dal::backend::kernel_dispatcher<backend::infer_kernel_cpu<Float, Method, Task>,
-                                            backend::infer_kernel_gpu<Float, Method, Task>>;
+            dal::backend::kernel_dispatcher<backend::infer_kernel_cpu<Float, Task>,
+                                            backend::infer_kernel_gpu<Float, Task>>;
         return kernel_dispatcher_t{}(ctx, params, input);
     }
 };
